refactor: moves layout components into layout folders
