+++ src/layouts/main-layout/MainLayout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router"
 import { SidebarProvider} from "@/components/ui/sidebar"
-import { DashboardSidebar } from "@/components/DashboardSidebar"
-import Header from "@/components/header/Header"
+import { DashboardSidebar } from "@/layouts/main-layout/DashboardSidebar"
+import Header from "@/layouts/main-layout/Header"
 
 export default function MainLayout() {
     return (
