+++ src/pages/auth/sign-up/index.tsx
@@ -1,5 +1,5 @@
 import SignUpForm from "@/features/auth/components/SignUpForm";
-import { AuthHeader } from "@/features/auth/components/AuthHeader";
+import { AuthHeader } from "@/layouts/auth-layout/AuthHeader";
 import { OAuthButtons } from "@/features/auth/components/OauthButtons";
 
 export default function SignUpPage() {
